added optional shadow topic when moving a topic fixed deletion of forums

+++ pnForum/pnadminapi.php
@@ -185,20 +185,6 @@ function pnForum_adminapi_deletecategory($args)
 	    $pntable =& pnDBGetTables();
     	$cattable = $pntable['pnforum_categories'];
     	$catcolumn = $pntable['pnforum_categories_column'];
-    	$posttable = $pntable['pnforum_posts'];
-    	$postcolumn = $pntable['pnforum_posts_column'];
-    	$posttexttable = $pntable['pnforum_posts_text'];
-    	$posttextcolumn = $pntable['pnforum_posts_text_column'];
-    	$cattable = $pntable['pnforum_categories'];
-    	$catcolumn = $pntable['pnforum_categories_column'];
-    	$forumtable = $pntable['pnforum_forums'];
-    	$forumcolumn = $pntable['pnforum_forums_column'];
-    	$topictable = $pntable['pnforum_topics'];
-    	$topiccolumn = $pntable['pnforum_topics_column'];
-    	$topicsubtable = $pntable['pnforum_topic_subscription'];
-    	$topicsubcolumn = $pntable['pnforum_topic_subscription_column'];
-    	$subscriptiontable = $pntable['pnforum_subscription'];
-    	$subscriptioncolumn = $pntable['pnforum_subscription_column'];
 /*
 	    // Confirm authorisation code
     	if (!pnSecConfirmAuthKey()) {
@@ -213,82 +199,8 @@ function pnForum_adminapi_deletecategory($args)
         if(is_array($forums) && count($forums)>0) { 
             foreach($forums as $forum) {
                 // remove all forums in this category
-		        $sql = "DELETE FROM $forumtable 
-		        		WHERE $forumcolumn[forum_id] = ".$forum['forum_id']."";
-		        $result = $dbconn->Execute($sql);
-		        if ($dbconn->ErrorNo() != 0) {
-		        	return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
-		        }
-		        $result->Close();
-                // delete forum subscription
-                $sql = "DELETE FROM $subscriptiontable
-                		WHERE $subscriptioncolumn[forum_id] = ".$forum['forum_id']."";
-                $result = $dbconn->Execute($sql);
-                if ($dbconn->ErrorNo() != 0) {
-                	return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
-                }
-		        $result->Close();
-                // get all topis in this forum by topic id
-		        $sql = "SELECT $topiccolumn[topic_id] 
-		        		FROM $topictable 
-		        		WHERE $topiccolumn[forum_id] = ".$forum['forum_id']."";
-		        $result = $dbconn->Execute($sql);
-		        if ($dbconn->ErrorNo() != 0) {
-		        	return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
-		        }
-                if($result->RecordCount()>0) {
-                    for (; !$result->EOF; $result->MoveNext()) {
-		        	    list($topic_id) = $result->fields;
-		        	    // delete topic
-		        	    $sql = "DELETE FROM $topictable
-		        	    		WHERE $topiccolumn[topic_id] = $topic_id";
-		        	    $result2 = $dbconn->Execute($sql);
-		        	    if ($dbconn->ErrorNo() != 0) {
-		        	    	return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
-		        	    }
-    		            $result2->Close();
-		        	    // delete topic subscription
-		        	    $sql = "DELETE FROM $topicsubtable
-		        	    		WHERE $topicsubcolumn[topic_id] = $topic_id";
-		        	    $result3 = $dbconn->Execute($sql);
-		        	    if ($dbconn->ErrorNo() != 0) {
-		        	    	return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
-		        	    }
-    		            $result3->Close();
-    		        }
-		        }
-		        $result->Close();
-               
-		        // get all posts in this forum by post_id
-		        $sql = "SELECT $postcolumn[post_id] 
-		        		FROM $posttable
-		        		WHERE $postcolumn[forum_id] = ".$forum['forum_id']."";
-		        $result = $dbconn->Execute($sql);
-		        if ($dbconn->ErrorNo() != 0) {
-		        	return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
-		        }
-                if($result->RecordCount()>0) {
-                    for (; !$result->EOF; $result->MoveNext()) {
-		        	    list($post_id) = $result->fields;
-		        	    // delete post
-		        	    $sql = "DELETE FROM $posttexttable
-		        	    		WHERE $posttextcolumn[post_id] = $post_id";
-		        	    $result2 = $dbconn->Execute($sql);
-		        	    if ($dbconn->ErrorNo() != 0) {
-		        	    	return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
-		        	    }
-    		            $result2->Close();
-		        	    // delete post text
-		        	    $sql = "DELETE FROM $posttable
-		        	    		WHERE $postcolumn[post_id] = $post_id";
-		        	    $result3 = $dbconn->Execute($sql);
-		        	    if ($dbconn->ErrorNo() != 0) {
-		        	    	return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
-		        	    }
-    		            $result3->Close();
-    		        }
-		        }
-		        $result->Close();
+                pnForum_userapi_deleteforum(array('forum_id' => $forum['forum_id'],
+                                                  'ok'       => 1));
 		    }  //foreach forum
 	    }
 	    // now we can delete the category
@@ -1168,21 +1080,27 @@ function pnForum_adminapi_editforum($args)
 
 /**
  * delete forum
+ *
+ *@params $args['forum_id']
+ *@params $args['ok']
+ *
  */
 function pnForum_adminapi_deleteforum($args)
 {
 // $forum_id, $ok
-	// global $hlpfile;
     extract($args);
     unset($args);
+
     if (!pnSecAuthAction(0, 'pnForum::', "::", ACCESS_ADMIN)) { 
-    	return showforumerror(_PNFORUM_NOAUTH, __FILE__, __LINE__); 
+    	return showforumerror(_PNFORUM_NOAUTH_TOADMIN, __FILE__, __LINE__); 
     }
 
 	pnModDBInfoLoad('pnForum');
 	$dbconn =& pnDBGetConn(true);
 	$pntable =& pnDBGetTables();
-	if($ok=="1") {
+	
+	if($ok==1) {
+        // delet forum
 		$sql = "DELETE FROM ".$pntable['pnforum_forums']." 
 				WHERE forum_id = '".pnVarPrepForStore($forum_id)."'";
 		$result = $dbconn->Execute($sql);
@@ -1196,7 +1114,44 @@ function pnForum_adminapi_deleteforum($args)
 		if ($dbconn->ErrorNo() != 0) {
     		return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
 		}
+        // delete forum subscription
+		$sql = "DELETE FROM ".$pntable['pnforum_subscription']." 
+				WHERE forum_id = '".pnVarPrepForStore($forum_id)."'";
+		$result = $dbconn->Execute($sql);
+		if ($dbconn->ErrorNo() != 0) {
+    		return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
+		}
 
+        // topics
+		$sql = "SELECT topic_id 
+				FROM ".$pntable['pnforum_topics']." 
+				WHERE forum_id = '".pnVarPrepForStore($forum_id)."'";
+		$result = $dbconn->Execute($sql);
+		if ($dbconn->ErrorNo() != 0) {
+    		return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
+		}
+        if($result->RecordCount()>0) {
+            for (; !$result->EOF; $result->MoveNext()) {
+    			list($topic_id) = $result->fields;
+	    		$sql = "DELETE FROM ".$pntable['pnforum_topic_subscription']." 
+		    			WHERE topic_id = '".pnVarPrepForStore($topic_id)."'";
+			    $del = $dbconn->Execute($sql);
+			    if ($dbconn->ErrorNo() != 0) {
+        		    return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
+			    }
+			    $del->Close();
+			    $sql = "DELETE FROM ".$pntable['pnforum_topics']." 
+				    	WHERE topic_id = '".pnVarPrepForStore($topic_id)."'";
+			    $del = $dbconn->Execute($sql);
+			    if ($dbconn->ErrorNo() != 0) {
+        		    return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
+			    }
+			    $del->Close();
+		    }
+		}
+        $result->Close();
+
+        // posts
 		$sql = "SELECT post_id 
 				FROM ".$pntable['pnforum_posts']." 
 				WHERE forum_id = '".pnVarPrepForStore($forum_id)."'";
@@ -1204,21 +1159,26 @@ function pnForum_adminapi_deleteforum($args)
 		if ($dbconn->ErrorNo() != 0) {
     		return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
 		}
-		while (!$result->EOF) {
-			list($post_id) = $result->fields;
-			$sql = "DELETE FROM ".$pntable['pnforum_posts_text']." 
-					WHERE post_id = '".pnVarPrepForStore($post_id)."'";
-			$del = $dbconn->Execute($sql);
-			if ($dbconn->ErrorNo() != 0) {
-        		return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
-			}
-			$sql = "DELETE FROM ".$pntable['pnforum_posts']." 
-					WHERE post_id = '".pnVarPrepForStore($post_id)."'";
-			$del = $dbconn->Execute($sql);
-			if ($dbconn->ErrorNo() != 0) {
-        		return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
-			}
+        if($result->RecordCount()>0) {
+            for (; !$result->EOF; $result->MoveNext()) {
+    			list($post_id) = $result->fields;
+	    		$sql = "DELETE FROM ".$pntable['pnforum_posts_text']." 
+		    			WHERE post_id = '".pnVarPrepForStore($post_id)."'";
+			    $del = $dbconn->Execute($sql);
+			    if ($dbconn->ErrorNo() != 0) {
+        		    return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
+			    }
+			    $del->Close();
+			    $sql = "DELETE FROM ".$pntable['pnforum_posts']." 
+				    	WHERE post_id = '".pnVarPrepForStore($post_id)."'";
+			    $del = $dbconn->Execute($sql);
+			    if ($dbconn->ErrorNo() != 0) {
+        		    return showforumsqlerror(_PNFORUM_ERROR_CONNECT,$sql,$dbconn->ErrorNo(),$dbconn->ErrorMsg(), __FILE__, __LINE__);
+			    }
+			    $del->Close();
+		    }
 		}
+        $result->Close();
 	}
     return;
 }
